feat: 공통 로직 분리
공통 로직 분리.

부모 클래스 필드들을 어떻게 초기화해야하는지 헷갈렸다.
해결 :
자식 클래스는 컴파일러가 super()를
자동으로 넣어주기 때문에 부모 생성자
를 잘 쓰면 된다
